ROutputParser as interface, bump major version
