+++ RCaller/build.gradle
@@ -8,7 +8,7 @@ repositories {
 }
 
 group = 'com.github.jbytecode'
-version = '3.1.0-SNAPSHOT'
+version = '4.0.0-SNAPSHOT'
 description = 'RCaller is a software library which simplifies performing data analysis and statistical calculations in Java using R. The details are hidden from users including transferring data between platforms, function calls, and retrieving results.'
 
 dependencies {
@@ -16,6 +16,7 @@ dependencies {
     compileOnly 'org.apache.arrow:arrow-vector:[3.0.0,)'
 
     testImplementation 'junit:junit:4.13.1'
+    testImplementation 'org.apache.arrow:arrow-vector:[3.0.0,)'
     testImplementation 'org.apache.arrow:arrow-memory-netty:[3.0.0,)'
 }
